Add API list view
