+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Apis.jsx
@@ -19,7 +19,7 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 
-import Listing from './Listing';
+import Listing from './Listing/Listing';
 import Details from './Details';
 import ApiCreate from './Create/ApiCreate';
 import { PageNotFound } from '../Base/Errors';
