+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Listing/Listing.jsx
@@ -18,79 +18,25 @@
 
 import React from 'react';
 import qs from 'qs';
-
-import { Link } from 'react-router-dom';
-import { Col, Menu, Row, Table } from 'antd';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
 import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import IconButton from '@material-ui/core/IconButton';
-import List from '@material-ui/icons/List';
-import GridIcon from '@material-ui/icons/GridOn';
-import APICreateMenu from './components/APICreateMenu';
-import Alert from '../../Shared/Alert';
+import Grid from '@material-ui/core/Grid';
 
-import { ScopeValidation, resourceMethod, resourcePath } from '../../../data/ScopeValidation';
-import ApiThumb from './components/ApiThumb';
+import Alert from '../../Shared/Alert';
+import PageNavigation from '../APIsNavigation';
+import PageContainer from '../../Base/container/';
 import API from '../../../data/api.js';
 import { Progress } from '../../Shared';
 import ResourceNotFound from '../../Base/Errors/ResourceNotFound';
 import SampleAPI from './SampleAPI/SampleAPI';
-
-const styles = theme => ({
-    rightIcon: {
-        marginLeft: theme.spacing.unit,
-    },
-    button: {
-        margin: theme.spacing.unit,
-        marginBottom: 20,
-    },
-    titleBar: {
-        display: 'flex',
-        justifyContent: 'space-between',
-        borderBottomWidth: '1px',
-        borderBottomStyle: 'solid',
-        borderColor: theme.palette.text.secondary,
-        marginBottom: 20,
-    },
-    buttonLeft: {
-        alignSelf: 'flex-start',
-        display: 'flex',
-    },
-    buttonRight: {
-        alignSelf: 'flex-end',
-        display: 'flex',
-    },
-    title: {
-        display: 'inline-block',
-        marginRight: 50,
-    },
-    addButton: {
-        display: 'inline-block',
-        marginBottom: 20,
-        zIndex: 1,
-    },
-    popperClose: {
-        pointerEvents: 'none',
-    },
-});
-const menu = (
-    <Menu>
-        <Link to='/api/create/swagger'>
-            <Menu.Item>Create new API with Swagger</Menu.Item>
-        </Link>
-        <Link to='/api/create/rest'>
-            <Menu.Item>Create new API</Menu.Item>
-        </Link>
-    </Menu>
-);
+import CardView from './CardView/CardView';
+// import TableView from './TableView/TableView';
+import TopMenu from './components/TopMenu';
 
 /**
  * Render the APIs Listing page, This is the Default Publisher Landing page as well
+ *
  * @class Listing
- * @extends {React.Component} @inheritdoc
+ * @extends {React.Component}
  */
 class Listing extends React.Component {
     /**
@@ -100,9 +46,10 @@ class Listing extends React.Component {
      */
     constructor(props) {
         super(props);
-        this.state = { listType: 'grid', apis: null };
+        this.state = { isCardView: true, apis: null };
         this.handleApiDelete = this.handleApiDelete.bind(this);
         this.updateApi = this.updateApi.bind(this);
+        this.toggleView = this.toggleView.bind(this);
     }
 
     /**
@@ -128,9 +75,14 @@ class Listing extends React.Component {
             });
     }
 
-    setListType = (value) => {
-        this.setState({ listType: value });
-    };
+    /**
+     *
+     *
+     * @memberof Listing
+     */
+    toggleView() {
+        this.setState({ isCardView: !this.state.isCardView });
+    }
 
     /**
      * Update Sample API
@@ -185,110 +137,37 @@ class Listing extends React.Component {
      * @memberof Listing
      */
     render() {
-        const { classes } = this.props;
-        const { apis } = this.state;
-        if (this.state.notFound) {
-            return <ResourceNotFound />;
-        }
-        const columns = [
-            {
-                title: 'Name',
-                dataIndex: 'name',
-                key: 'name',
-                render: (text, record) => <Link to={'/apis/' + record.id}>{text}</Link>,
-            },
-            {
-                title: 'Context',
-                dataIndex: 'context',
-                key: 'context',
-            },
-            {
-                title: 'Version',
-                dataIndex: 'version',
-                key: 'version',
-            },
-            {
-                title: 'Action',
-                key: 'action',
-                render: record => (
-                    <ScopeValidation resourcePath={resourcePath.SINGLE_API} resourceMethod={resourceMethod.DELETE}>
-                        <Button
-                            style={{ fontSize: 10, padding: '0px', margin: '0px' }}
-                            color='primary'
-                            onClick={() => this.handleApiDelete(record.id, record.name)}
-                        >
-                            Delete
-                        </Button>
-                    </ScopeValidation>
-                ),
-            },
-        ];
-        if (!apis) {
-            return <Progress />;
-        } else if (apis.count === 0) {
-            return <SampleAPI />;
-        } else {
+        const { apis, notFound, isCardView } = this.state;
+        if (notFound) {
             return (
-                <Grid container spacing={0} justify='center'>
-                    <Grid item md={12} className={classes.titleBar}>
-                        <div className={classes.buttonLeft}>
-                            <div className={classes.title}>
-                                <Typography variant='display2' gutterBottom>
-                                    APIs
-                                </Typography>
-                            </div>
-                            <APICreateMenu buttonProps={{ size: 'small', color: 'primary', variant: 'outlined' }}>
-                                Create API
-                            </APICreateMenu>
-                        </div>
-                        <div className={classes.buttonRight}>
-                            <IconButton
-                                className={classes.button}
-                                aria-label='Delete'
-                                onClick={() => this.setListType('list')}
-                            >
-                                <List />
-                            </IconButton>
-                            <IconButton
-                                className={classes.button}
-                                aria-label='Delete'
-                                onClick={() => this.setListType('grid')}
-                            >
-                                <GridIcon />
-                            </IconButton>
-                        </div>
-                    </Grid>
-
-                    <Grid item xs={12}>
-                        {this.state.listType === 'list' ? (
-                            <Row type='flex' justify='start'>
-                                <Col span={24}>
-                                    <Table
-                                        columns={columns}
-                                        dataSource={this.state.apis.list}
-                                        bordered
-                                        locale={{ emptyText: 'There is no data to display' }}
-                                    />
-                                </Col>
-                            </Row>
-                        ) : (
-                            <Grid container spacing={8}>
-                                {this.state.apis.list.map((api, i) => {
-                                    return (
-                                        <ApiThumb
-                                            key={api.id}
-                                            listType={this.state.listType}
-                                            api={api}
-                                            updateApi={this.updateApi}
-                                        />
-                                    );
-                                })}
-                            </Grid>
-                        )}
-                    </Grid>
-                </Grid>
+                <PageContainer pageNav={<PageNavigation />}>
+                    <ResourceNotFound />
+                </PageContainer>
             );
         }
+        if (!apis) {
+            return (
+                <PageContainer pageNav={<PageNavigation />}>
+                    <Progress />
+                </PageContainer>
+            );
+        }
+        if (apis.count === 0) {
+            return (
+                <PageContainer pageNav={<PageNavigation />}>
+                    <SampleAPI />
+                </PageContainer>
+            );
+        }
+
+        return (
+            <PageContainer
+                pageTopMenu={<TopMenu toggleView={this.toggleView} isCardView={isCardView} />}
+                pageNav={<PageNavigation />}
+            >
+                {isCardView ? <CardView apis={apis} /> : 'Table view here'}
+            </PageContainer>
+        );
     }
 }
 
@@ -297,4 +176,4 @@ Listing.propTypes = {
     theme: PropTypes.shape({}).isRequired,
 };
 
-export default withStyles(styles, { withTheme: true })(Listing);
+export default Listing;
