--- features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Listing/index.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import PageNavigation from '../APIsNavigation';
-import PageContainer from '../../Base/container/';
-import Listing from './Listing';
-
-/**
- * Export the composed APIs Listing page
- * @returns {React.Component} @inheritdoc
- */
-export default () => {
-    return (
-        <PageContainer pageNav={<PageNavigation />}>
-            <Listing />
-        </PageContainer>
-    );
-};
