+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Base/container/index.jsx
@@ -38,7 +38,7 @@ const Container = (props) => {
                 {pageNav}
             </Grid>
             <Grid item className={classes.pageContainer} xs={10} sm={11} md={10} lg={11}>
-                <Grid container>
+                <Grid className={classes.pageContainer} container>
                     {pageTopMenu && (
                         <Grid className={classes.pageTopMenu} item xs={12} sm={12} md={12} lg={12}>
                             {pageTopMenu}
